refactor: deduplicate schemas and package api
