+++ api/__init__.py
@@ -1,0 +1,1 @@
+"""Doc_GPT API package."""
