--- api/init.py
@@ -1,2 +1,0 @@
-# D:\projects\Doc_GPT\api\__init__.py
-# (empty)
