+++ requirements.txt
@@ -23,3 +23,4 @@ rank-bm25==0.2.2
 tqdm==4.66.5
 tenacity==9.0.0
 jinja2==3.1.4
+openai==1.52.0
